+++ src/App.tsx
@@ -2,14 +2,11 @@ import { useState, useEffect } from 'react';
 import './App.scss';
 import { MoviesList } from './components/MoviesList';
 import { FindMovie } from './components/FindMovie';
-import { getMovie } from './api';
-import { MovieData } from './types/MovieData';
 import { Movie } from './types/Movie';
 
 export const App = () => {
   const [movies, setMovies] = useState<Movie[]>([]);
   const [query, setQuery] = useState<string>('');
-  const [querySearch, setQuerySearch] = useState<string>('');
   const [movie, setMovie] = useState<Movie | null>(null);
   const [searchOk, setSearchOk] = useState(true);
   const [searchEmpty, setSearchEmpty] = useState(true);
@@ -21,38 +18,9 @@ export const App = () => {
     if (!isDuplicate) {
       setMovies(prevMovies => [...prevMovies, newMovie]);
     }
-
     setQuery('');
     setMovie(null);
   };
-
-  function transformMovieData(data: MovieData): Movie {
-    return {
-      title: data.Title || '',
-      description: data.Plot || '',
-      imgUrl:
-        data.Poster !== 'N/A'
-          ? data.Poster
-          : 'https://via.placeholder.com/360x270.png?text=no%20preview',
-      imdbUrl: `https://www.imdb.com/title/${data.imdbID}`,
-      imdbId: data.imdbID || '',
-    };
-  }
-
-  useEffect(() => {
-    if (querySearch) {
-      getMovie(querySearch).then(response => {
-        if ('Title' in response) {
-          setMovie(transformMovieData(response));
-          setSearchOk(true);
-          setLoading(false);
-        } else {
-          setLoading(false);
-          setSearchOk(false);
-        }
-      });
-    }
-  }, [querySearch]);
 
   return (
     <div className="page">
@@ -66,7 +34,6 @@ export const App = () => {
           query={query}
           movies={movies}
           setQuery={setQuery}
-          setQuerySerch={setQuerySearch}
           onAdd={addMovie}
           searchOk={searchOk}
           setSearchOk={setSearchOk}
@@ -74,6 +41,7 @@ export const App = () => {
           setSearchEmpty={setSearchEmpty}
           loading={loading}
           setLoading={setLoading}
+          setMovie={setMovie}
         />
       </div>
     </div>
